requirements.txt
torchvision==0.7.0+cu101
scipy==1.1.0
numpy==1.19.2
opencv_python==4.5.1.48
scikit_image==0.18.1
matplotlib==3.3.2
torch==1.6.0+cu101
Pillow==8.3.2
skimage==0.0
tensorboardX==2.4
